+++ vortex-btrblocks/src/lib.rs
@@ -78,8 +78,8 @@ pub use builder::BtrBlocksCompressorBuilder;
 pub use canonical_compressor::BtrBlocksCompressor;
 pub use schemes::patches::compress_patches;
 pub use vortex_compressor::CascadingCompressor;
-pub use vortex_compressor::ctx::CompressorContext;
-pub use vortex_compressor::ctx::MAX_CASCADE;
+pub use vortex_compressor::scheme::CompressorContext;
+pub use vortex_compressor::scheme::MAX_CASCADE;
 pub use vortex_compressor::scheme::Scheme;
 pub use vortex_compressor::scheme::SchemeExt;
 pub use vortex_compressor::scheme::SchemeId;
