+++ vortex-compressor/src/compressor/constant.rs
