+++ vortex-compressor/src/lib.rs
@@ -63,14 +63,10 @@
 //! with a short `jq` query.
 
 pub mod builtins;
-pub mod ctx;
-pub mod estimate;
 pub mod scheme;
 pub mod stats;
 
-mod constant;
-mod sample;
-
 mod compressor;
-mod trace;
 pub use compressor::CascadingCompressor;
+
+mod trace;
