+++ vortex-compressor/src/scheme/estimate.rs
@@ -1,25 +1,16 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
-//! Compression ratio estimation types and sampling-based estimation.
+//! Compression ratio estimation types returned by schemes.
 
 use std::fmt;
 
-use vortex_array::ArrayRef;
-use vortex_array::Canonical;
 use vortex_array::ExecutionCtx;
-use vortex_array::IntoArray;
 use vortex_error::VortexResult;
 
 use crate::CascadingCompressor;
-use crate::ctx::CompressorContext;
-use crate::sample::SAMPLE_SIZE;
-use crate::sample::sample;
-use crate::sample::sample_count_approx_one_percent;
-use crate::scheme::Scheme;
-use crate::scheme::SchemeExt;
+use crate::scheme::CompressorContext;
 use crate::stats::ArrayAndStats;
-use crate::trace;
 
 /// Closure type for [`DeferredEstimate::Callback`].
 ///
@@ -44,10 +35,12 @@ pub type EstimateFn = dyn FnOnce(
     + Send
     + Sync;
 
-/// The result of a [`Scheme`]'s compression ratio estimation.
+/// The result of a [`Scheme`](crate::scheme::Scheme)'s compression ratio estimation.
 ///
-/// This type is returned by [`Scheme::expected_compression_ratio`] to tell the compressor how
-/// promising this scheme is for a given array without performing any expensive work.
+/// This type is returned by
+/// [`Scheme::expected_compression_ratio`](crate::scheme::Scheme::expected_compression_ratio) to
+/// tell the compressor how promising this scheme is for a given array without performing any
+/// expensive work.
 ///
 /// [`CompressionEstimate::Verdict`] means the scheme already knows the terminal answer.
 /// [`CompressionEstimate::Deferred`] means the compressor must do extra work before the scheme can
@@ -119,7 +112,7 @@ pub enum EstimateScore {
 
 impl EstimateScore {
     /// Converts measured sample sizes into a ranked estimate.
-    pub(super) fn from_sample_sizes(before_nbytes: u64, after_nbytes: u64) -> Self {
+    pub(crate) fn from_sample_sizes(before_nbytes: u64, after_nbytes: u64) -> Self {
         if after_nbytes == 0 {
             Self::ZeroBytes
         } else {
@@ -139,7 +132,7 @@ impl EstimateScore {
     }
 
     /// Returns whether this estimate is eligible to compete.
-    fn is_valid(self) -> bool {
+    pub(crate) fn is_valid(self) -> bool {
         match self {
             Self::FiniteCompression(ratio) => {
                 ratio.is_finite() && !ratio.is_subnormal() && ratio > 1.0
@@ -149,7 +142,7 @@ impl EstimateScore {
     }
 
     /// Returns whether this estimate beats another valid estimate.
-    fn beats(self, other: Self) -> bool {
+    pub(crate) fn beats(self, other: Self) -> bool {
         match (self, other) {
             (Self::ZeroBytes, _) => false,
             (Self::FiniteCompression(_), Self::ZeroBytes) => true,
@@ -158,81 +151,6 @@ impl EstimateScore {
             }
         }
     }
-}
-
-/// Winner estimate carried from scheme selection into result tracing.
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub(super) enum WinnerEstimate {
-    /// The scheme must be used immediately.
-    AlwaysUse,
-    /// The scheme won by a ranked estimate.
-    Score(EstimateScore),
-}
-
-impl WinnerEstimate {
-    /// Returns the traceable numeric ratio for the winning estimate.
-    pub(super) fn trace_ratio(self) -> Option<f64> {
-        match self {
-            Self::AlwaysUse => None,
-            Self::Score(score) => score.finite_ratio(),
-        }
-    }
-}
-
-/// Returns `true` if `score` beats the current best estimate.
-pub(super) fn is_better_score(
-    score: EstimateScore,
-    best: Option<&(&'static dyn Scheme, EstimateScore)>,
-) -> bool {
-    score.is_valid() && best.is_none_or(|(_, best_score)| score.beats(*best_score))
-}
-
-/// Estimates compression ratio by compressing a ~1% sample of the data.
-///
-/// Creates a new [`ArrayAndStats`] for the sample so that stats are generated from the sample, not
-/// the full array.
-///
-/// # Errors
-///
-/// Returns an error if sample compression fails.
-pub(super) fn estimate_compression_ratio_with_sampling<S: Scheme + ?Sized>(
-    compressor: &CascadingCompressor,
-    scheme: &S,
-    array: &ArrayRef,
-    compress_ctx: CompressorContext,
-    exec_ctx: &mut ExecutionCtx,
-) -> VortexResult<EstimateScore> {
-    let sample_array = if compress_ctx.is_sample() {
-        array.clone()
-    } else {
-        let sample_count = sample_count_approx_one_percent(array.len());
-        // `ArrayAndStats` expects a canonical array (so that it can easily compute lazy stats).
-        let canonical: Canonical = sample(array, SAMPLE_SIZE, sample_count).execute(exec_ctx)?;
-        canonical.into_array()
-    };
-
-    let sample_data = ArrayAndStats::new(sample_array, scheme.stats_options());
-    let error_ctx = trace::enabled_error_context(&compress_ctx);
-    let sample_ctx = compress_ctx.with_sampling();
-
-    let compressed = match scheme.compress(compressor, &sample_data, sample_ctx, exec_ctx) {
-        Ok(compressed) => compressed,
-        Err(err) => {
-            trace::sample_compress_failed(scheme.id(), error_ctx.as_ref(), &err);
-            return Err(err);
-        }
-    };
-
-    let after = compressed.nbytes();
-    let before = sample_data.array().nbytes();
-
-    let score = EstimateScore::from_sample_sizes(before, after);
-
-    if matches!(score, EstimateScore::ZeroBytes) {
-        trace::zero_byte_sample_result(scheme.id(), before);
-    }
-
-    Ok(score)
 }
 
 impl fmt::Debug for DeferredEstimate {
